refactor: reorganize package
Change package to com.redhat.console.integrations,
and move Splunk specific stuff to .splunk subpackage.

EVNT-601
