+++ splunk-quarkus/src/main/java/com/redhat/console/integrations/EventAppender.java
@@ -1,4 +1,4 @@
-package com.redhat.console.notifications.splunkintegration;
+package com.redhat.console.integrations;
 
 import org.apache.camel.AggregationStrategy;
 import org.apache.camel.Exchange;
