+++ splunk-quarkus/src/main/java/com/redhat/console/integrations/EventPicker.java
@@ -1,4 +1,4 @@
-package com.redhat.console.notifications.splunkintegration;
+package com.redhat.console.integrations;
 
 import org.apache.camel.Exchange;
 import org.apache.camel.Message;
